\name{ArrayBlock-class}
\docType{class}

\alias{class:ArrayBlock}
\alias{ArrayBlock-class}
\alias{ArrayBlock}

\alias{dim,ArrayBlock-method}
\alias{length,ArrayBlock-method}
\alias{show,ArrayBlock-method}

\alias{makeNindexFromArrayBlock}

\title{ArrayBlock objects}

\description{
  ArrayBlock objects are used internally to support block processing of
  array-like objects.
}

\keyword{internal}
