\name{ArrayBlocks-class}
\docType{class}

\alias{class:ArrayBlocks}
\alias{ArrayBlocks-class}
\alias{ArrayBlocks}

\alias{length,ArrayBlocks-method}
\alias{show,ArrayBlocks-method}

\title{ArrayBlocks objects}

\description{
  ArrayBlocks objects are used internally to support block processing of
  array-like objects.
}

\keyword{internal}
