\name{block_processing}

\alias{block_processing}

\alias{class:ArrayBlocks}
\alias{ArrayBlocks-class}
\alias{ArrayBlocks}

\alias{class:OnDiskArrayDump}
\alias{OnDiskArrayDump-class}
\alias{OnDiskArrayDump}

\alias{write_to_dump}
\alias{close,OnDiskArrayDump-method}

\title{Internal utilities for block processing an array}

\description{
  Internal utilities for block processing an array.
}

\keyword{internal}
