\name{ConformableSeedCombiner-class}

\alias{dim,ConformableSeedCombiner-method}
\alias{dimnames,ConformableSeedCombiner-method}
\alias{extract_array,ConformableSeedCombiner-method}

\alias{path,ConformableSeedCombiner-method}
\alias{path<-,ConformableSeedCombiner-method}

\title{ConformableSeedCombiner objects}

\description{
  ConformableSeedCombiner objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
