\name{DelayedVariadicIsoOp-class}

\alias{dim,DelayedVariadicIsoOp-method}
\alias{dimnames,DelayedVariadicIsoOp-method}
\alias{extract_array,DelayedVariadicIsoOp-method}

\alias{path,DelayedVariadicIsoOp-method}
\alias{path<-,DelayedVariadicIsoOp-method}

\title{DelayedVariadicIsoOp objects}

\description{
  DelayedVariadicIsoOp objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
