\name{DelayedVariadicOp-class}

\alias{dim,DelayedVariadicOp-method}
\alias{dimnames,DelayedVariadicOp-method}
\alias{extract_array,DelayedVariadicOp-method}

\alias{path,DelayedVariadicOp-method}
\alias{path<-,DelayedVariadicOp-method}

\title{DelayedVariadicOp objects}

\description{
  DelayedVariadicOp objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
