\name{dump-utils}

\alias{dump-utils}
\alias{dump}

\alias{supportedRealizeBackends}
\alias{setRealizeBackend}
\alias{getRealizeBackend}

\alias{realize}

\alias{class:OnDiskArrayDump}
\alias{OnDiskArrayDump-class}
\alias{OnDiskArrayDump}

\alias{write_to_dump}
\alias{close,OnDiskArrayDump-method}

\title{Dump management}

\description{
  
  Internal utilities for block processing an array.
}

\keyword{methods}
