\name{GridOfBlocks-class}
\docType{class}

\alias{class:ArrayBlock}
\alias{ArrayBlock-class}
\alias{ArrayBlock}

\alias{dim,ArrayBlock-method}
\alias{length,ArrayBlock-method}
\alias{show,ArrayBlock-method}

\alias{makeNindexFromArrayBlock}

\alias{class:GridOfBlocks}
\alias{GridOfBlocks-class}
\alias{GridOfBlocks}

\alias{length,GridOfBlocks-method}
\alias{show,GridOfBlocks-method}

\title{GridOfBlocks and ArrayBlock objects}

\description{
  GridOfBlocks and ArrayBlock objects are used internally to support block
  processing of array-like objects.
}

\keyword{internal}
