\name{RealizationSink-class}
\docType{class}

\alias{class:RealizationSink}
\alias{RealizationSink-class}

\alias{write_to_sink}
\alias{close,RealizationSink-method}

\alias{supportedRealizeBackends}
\alias{setRealizeBackend}
\alias{getRealizeBackend}

\alias{RealizationSink}

\title{RealizationSink objects}

\description{
  RealizationSink objects are used behind the scene to realize DelayedArray
  objects on disk.
}

\details{
  Coming soon...
}

\keyword{methods}
\keyword{classes}
