\name{SeedBinder-class}

\alias{dim,SeedBinder-method}
\alias{dimnames,SeedBinder-method}
\alias{extract_array,SeedBinder-method}

\alias{path,SeedBinder-method}
\alias{path<-,SeedBinder-method}

\title{SeedBinder objects}

\description{
  SeedBinder objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
