\name{SeedDimCombiner-class}

\alias{dim,SeedDimCombiner-method}
\alias{dimnames,SeedDimCombiner-method}
\alias{extract_array,SeedDimCombiner-method}

\alias{aperm.SeedDimCombiner}
\alias{aperm,SeedDimCombiner-method}

\title{SeedDimCombiner objects}

\description{
  SeedDimCombiner objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
