\name{SeedDimPicker-class}

\alias{dim,SeedDimPicker-method}
\alias{dimnames,SeedDimPicker-method}
\alias{extract_array,SeedDimPicker-method}

\alias{aperm.SeedDimPicker}
\alias{aperm,SeedDimPicker-method}

\alias{seed,SeedDimPicker-method}
\alias{seed<-,SeedDimPicker-method}

\alias{path,SeedDimPicker-method}
\alias{path<-,SeedDimPicker-method}

\title{SeedDimPicker objects}

\description{
  SeedDimPicker objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
