\name{ArrayBlock-class}
\docType{class}

\alias{class:ArrayBlock}
\alias{ArrayBlock-class}
\alias{ArrayBlock}

\alias{dim,ArrayBlock-method}
\alias{length,ArrayBlock-method}
\alias{show,ArrayBlock-method}

\alias{makeNindexFromArrayBlock}

\alias{class:BlockGrid}
\alias{BlockGrid-class}
\alias{BlockGrid}

\alias{length,BlockGrid-method}
\alias{show,BlockGrid-method}

\title{ArrayBlock and BlockGrid objects}

\description{
  ArrayBlock and BlockGrid objects are used internally to support block
  processing of array-like objects.
}

\keyword{internal}
