\name{block_processing}

\alias{block_processing}

\alias{write_array_to_sink}

\title{Block processing of an array}

\description{
  A set of utilities for processing an array-like object block by block.
}

\details{
  Coming soon...
}

\seealso{
  \itemize{
    \item \link{DelayedArray} objects.

    \item \code{\link{realize}} for realizing a DelayedArray object in memory
          or on disk.
  }
}

\keyword{methods}
