\name{block_processing}

\alias{block_processing}

\alias{class:ArrayBlocks}
\alias{ArrayBlocks-class}
\alias{ArrayBlocks}

\alias{class:ArrayOnDiskDump}
\alias{ArrayOnDiskDump-class}
\alias{ArrayOnDiskDump}

\alias{write_to_dump}
\alias{write_to_dump,ANY,ArrayOnDiskDump-method}
\alias{write_to_dump,DelayedArray,ArrayOnDiskDump-method}

\title{Internal utilities for block processing an array}

\description{
  Internal utilities for block processing an array.
}

\keyword{internal}
