\name{block_processing}

\alias{block_processing}

\alias{class:ArrayBlocks}
\alias{ArrayBlocks-class}
\alias{ArrayBlocks}

\alias{getListElement,ArrayBlocks-method}
\alias{show,ArrayBlocks-method}

\title{Internal utilities for block processing an array}

\description{
  Internal utilities for block processing an array.
}

\keyword{internal}
