\name{DelayedAperm-class}

\alias{dim,DelayedAperm-method}
\alias{dimnames,DelayedAperm-method}
\alias{extract_array,DelayedAperm-method}

\title{DelayedAperm objects}

\description{
  DelayedAperm objects are used inside \link{DelayedArray} objects
  and are not intended to be exposed to the end user.
}

\seealso{
  \link{DelayedArray} objects.
}

\keyword{internal}
