\name{dump-utils}

\alias{dump-utils}
\alias{dump}

\alias{class:OnDiskArrayDump}
\alias{OnDiskArrayDump-class}
\alias{OnDiskArrayDump}

\alias{write_to_dump}
\alias{close,OnDiskArrayDump-method}

\title{Dump management utilities}

\description{
  OnDiskArrayDump objects are used behind the scene to realize DelayedArray
  object on disk.
}

\details{
  Coming soon...
}

\keyword{methods}
\keyword{classes}
