\name{RealizationSink-class}
\docType{class}

\alias{class:RealizationSink}
\alias{RealizationSink-class}

\alias{write_to_sink}
\alias{write_to_sink,DelayedArray,RealizationSink-method}
\alias{write_to_sink,ANY,RealizationSink-method}

\alias{close,RealizationSink-method}

\alias{class:arrayRealizationSink}
\alias{arrayRealizationSink-class}
\alias{arrayRealizationSink}

\alias{write_to_sink,array,arrayRealizationSink-method}
\alias{coerce,arrayRealizationSink,DelayedArray-method}

\alias{RealizationSink}

\title{RealizationSink objects}

\description{
  RealizationSink objects are used internally to control realization of
  DelayedArray objects in memory or on disk.
}

\keyword{internal}
